Changes to column types to allow non-integer intensities in metabolomics. Removed some rounding in metabolomics zscore calc, rounding removes information. Removed a case checking on <= 0 intensity value for log calculation, these values are removed in an earlier step. If they are not it is correct to throw an error rather than adding a small value to the intensity.

+++ ddl/postgres/tm_wz/wt_subject_metabolomics_logs.sql
@@ -3,7 +3,7 @@
 --
 CREATE TABLE wt_subject_metabolomics_logs (
     probeset character varying(500),
-    intensity_value bigint,
+    intensity_value double precision,
     pvalue double precision,
     num_calls bigint,
     assay_id bigint,
@@ -12,7 +12,7 @@ CREATE TABLE wt_subject_metabolomics_logs (
     subject_id character varying(50),
     trial_name character varying(50),
     timepoint character varying(100),
-    log_intensity bigint
+    log_intensity double precision
 );
 
 --
